feat: rename class name
